requirements.txt
bokeh==2.4.3
streamlit-bokeh-events==0.1.2
jinja2==3.1.0
gtts
